+++ blaze-core/src/test/java/com/fizzed/blaze/system/ExecCaptureOutputTest.java
@@ -15,119 +15,48 @@
  */
 package com.fizzed.blaze.system;
 
-import com.fizzed.blaze.core.ExecutableNotFoundException;
-import com.fizzed.blaze.Config;
-import com.fizzed.blaze.internal.ContextImpl;
+import com.fizzed.blaze.core.DefaultContext;
 import static com.fizzed.blaze.system.ShellTestHelper.getBinDirAsResource;
-import com.fizzed.blaze.internal.ConfigHelper;
 import com.fizzed.blaze.util.CaptureOutput;
-import com.fizzed.blaze.util.StreamableInput;
-import com.fizzed.blaze.util.StreamableOutput;
 import com.fizzed.blaze.util.Streamables;
-import java.io.ByteArrayInputStream;
-import java.io.ByteArrayOutputStream;
-import java.io.InputStream;
-import java.io.OutputStream;
-import java.nio.file.Paths;
-import java.util.Objects;
-import org.junit.Before;
 import org.junit.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.CoreMatchers.sameInstance;
 import static org.junit.Assert.assertThat;
-import static org.junit.Assume.assumeTrue;
-import static org.mockito.Mockito.atLeast;
-import static org.mockito.Mockito.spy;
-import static org.mockito.Mockito.verify;
 
-public class ExecTest {
-    private static final Logger log = LoggerFactory.getLogger(ExecTest.class);
-    
-    Config config;
-    ContextImpl context;
-    
-    @Before
-    public void setup() {
-        config = ConfigHelper.create(null);
-        context = spy(new ContextImpl(null, null, Paths.get("blaze.js"), config));
-    }
-    
-    @Before
-    public void doNotRunOnTravisCI() {
-        // NOTE: travis-ci has an issue w/ this unit test that never happens
-        // on a real system. Until an upstream issue with zt-exec is resolved
-        // we are going to give up trying to make this work on travis-ci
-        assumeTrue("Not running on travis-ci", !Objects.equals(System.getenv("TRAVIS"), "true"));
-    }
-    
-    @Test(expected=ExecutableNotFoundException.class)
-    public void notFind() throws Exception {
-        new Exec(context)
-            .command("thisdoesnotexist")
-            .run();
-    }
-    
+public class ExecCaptureOutputTest {
+    private static final Logger log = LoggerFactory.getLogger(ExecCaptureOutputTest.class);
+
     @Test
-    public void works() throws Exception {
-        Integer exitValue
-            = new Exec(context)
+    public void runCaptureOutput() throws Exception {
+        String output =
+            new Exec(new DefaultContext())
                 .command("hello-world-test")
                 .path(getBinDirAsResource())
-                .run();
-        
-        assertThat(exitValue, is(0));
-    }
-    
-    @Test
-    public void nullInputAndOutputs() throws Exception {
-        Integer exitValue
-            = new Exec(context)
-                .command("hello-world-test")
-                .path(getBinDirAsResource())
-                .pipeInput((StreamableInput)null)
-                .pipeOutput((StreamableOutput)null)
-                .pipeError((StreamableOutput)null)
-                .run();
-        
-        assertThat(exitValue, is(0));
-    }
-    
-    @Test
-    public void customInputAndOutputs() throws Exception {
-        // custom input and outputs
-        InputStream in = spy(new ByteArrayInputStream(new byte[0]));
-        OutputStream out = spy(new ByteArrayOutputStream());
-        OutputStream err = spy(new ByteArrayOutputStream());
-        
-        new Exec(context)
-            .command("hello-world-test")
-            .path(getBinDirAsResource())
-            .pipeInput(in)
-            .pipeOutput(out)
-            .pipeError(err)
-            .run();
-        
-        // verify all streams were closed
-        verify(in, atLeast(1)).close();
-        verify(out, atLeast(1)).close();
-        verify(err, atLeast(1)).close();
-    }
-    
-    @Test
-    public void captureOutput() throws Exception {
-        CaptureOutput capture = Streamables.captureOutput();
-        
-        new Exec(context)
-            .command("hello-world-test")
-            .path(getBinDirAsResource())
-            .pipeInput(Streamables.standardInput())
-            .pipeOutput(capture)
-            .run();
-            
-        String output = capture.asString();
+                .disablePipeInput()
+                .disablePipeError()
+                .runCaptureOutput()
+                .toString();
         
         assertThat(output.trim(), is("Hello World 7586930100"));
+    }
+    
+    @Test
+    public void runCaptureOutputUseExistingCaptureOutputIfSupplied() throws Exception {
+        CaptureOutput captureOutput = Streamables.captureOutput();
+        CaptureOutput output =
+            new Exec(new DefaultContext())
+                .command("hello-world-test")
+                .path(getBinDirAsResource())
+                .disablePipeInput()
+                .disablePipeError()
+                .pipeOutput(captureOutput)
+                .runCaptureOutput();
+        
+        assertThat(output, sameInstance(captureOutput));
+        assertThat(output.toString().trim(), is("Hello World 7586930100"));
     }
     
 }