+++ CHANGELOG.md
@@ -3,10 +3,11 @@ Blaze by Fizzed
 
 #### x.x.x
 
+ - Systems.exec and SecureShells.exec support a convenient `runCaptureOutput`
+   method to simplify the use-case of capturing the output of a process.
  - Convenience disablePipeInput, disablePipeOutput, and disablePipeError added
    for Systems.exec and SecureShells.exec to simplify syntax of disabling any
    of those particular inputs/outputs.
- - 
 
 #### 0.9.1 - 2016-04-08